make exceptions shared
